server/ocr_service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
keras-ocr==0.8.3
Pillow==10.0.0
pdf2image==1.16.3
python-multipart==0.0.6
numpy==1.25.2
typing-extensions==4.7.1